Rich text editing compose window with toolbar and formatting sub toolbar. New attachment list style.
Fixed: the compose xib instantiated a copy of its own controller inside of it even that the controller was what was loading the xib to begin with, leading to strange bugs.

Decoupled the representation of attachments to its own class, SMAttachment, where each attachment belongs to an SMEmail. Observable.

List of attachments in a collection view driving by an array controller, where each attachment has view, download and delete buttons.

Attachment list automatically grows to accommodate up to 3 attachments, before switching to a scrollable area.

Attachment button in the toolbar using an invisible overlay of an upload button.

Rich text editor is based on WKTextView and the formatting toolbar has its own controller.